+++ backend/src/main/java/com/woori/wonfit/log/loginlog/controller/MemberLoginLog.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.loginlog.controller;
+
+public class MemberLoginLog {
+}
