+++ backend/src/main/java/com/woori/wonfit/log/loginlog/dto/MemberLoginLogResponse.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.loginlog.dto;
+
+public class MemberLoginLogResponse {
+}
