+++ backend/src/main/java/com/woori/wonfit/log/loginlog/service/MemberLoginLogService.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.loginlog.service;
+
+public class MemberLoginLogService {
+}
