+++ backend/src/main/java/com/woori/wonfit/log/searchlog/controller/MemberSearchLogController.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.searchlog.controller;
+
+public class MemberSearchLogController {
+}
