+++ backend/src/main/java/com/woori/wonfit/log/searchlog/controller/SearchLogController.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.searchlog.controller;
+
+public class SearchLogController {
+}
