+++ backend/src/main/java/com/woori/wonfit/log/searchlog/service/MemberSearchLogService.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.searchlog.service;
+
+public class MemberSearchLogService {
+}
