+++ backend/src/main/java/com/woori/wonfit/log/subscribelog/dto/MemberSubLogResponse.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.log.subscribelog.dto;
+
+public class MemberSubLogResponse {
+}
