--- backend/src/main/java/com/woori/wonfit/member/loginlog/controller/MemberLoginLog.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.loginlog.controller;
-
-public class MemberLoginLog {
-}
