--- backend/src/main/java/com/woori/wonfit/member/loginlog/dto/MemberLoginLogResponse.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.loginlog.dto;
-
-public class MemberLoginLogResponse {
-}
