--- backend/src/main/java/com/woori/wonfit/member/loginlog/service/MemberLoginLogService.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.loginlog.service;
-
-public class MemberLoginLogService {
-}
