--- backend/src/main/java/com/woori/wonfit/member/searchlog/controller/MemberSearchLogController.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.searchlog.controller;
-
-public class MemberSearchLogController {
-}
