--- backend/src/main/java/com/woori/wonfit/member/searchlog/controller/SearchLogController.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.searchlog.controller;
-
-public class SearchLogController {
-}
