--- backend/src/main/java/com/woori/wonfit/member/searchlog/dto/MemberSearchLogResponse.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.searchlog.dto;
-
-public class MemberSearchLogResponse {
-}
