--- backend/src/main/java/com/woori/wonfit/member/searchlog/service/SearchLogService.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.searchlog.service;
-
-public class SearchLogService {
-}
