--- backend/src/main/java/com/woori/wonfit/member/subscribelog/controller/SubscribeLogController.java
@@ -1,25 +1,0 @@
-package com.woori.wonfit.member.subscribelog.controller;
-
-import com.woori.wonfit.member.subscribelog.domain.SubscribeLog;
-import com.woori.wonfit.member.subscribelog.service.SubscribeLogService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/sublogs")
-@RequiredArgsConstructor
-public class SubscribeLogController {
-
-    private final SubscribeLogService SubscribeLogService;
-
-    @GetMapping("/member/{memberId}")
-    public List<SubscribeLog> getSubscribeLogsByMemberId(@PathVariable Long memberId) {
-        return SubscribeLogService.SubscribeLogsByMemberId(memberId);
-
-    }
-}
