--- backend/src/main/java/com/woori/wonfit/member/subscribelog/dto/MemberSubLogResponse.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.subscribelog.dto;
-
-public class MemberSubLogResponse {
-}
